+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 * Initial implementation of Record.Checksum(), intended for easy and fast comparison of recordable objects.
+* Added Intent property to Recorder to indicate Serialization/Cloning/Checksum purpose.
 
 ### Improved
 * Exception messages for ConfigErrors() and PostLoad() now include the Dec name.